# Conflicts:
#	RoombaRampage/RoombaRampage/ECS/System/RenderSystem.cpp
#	RoombaRampage/RoombaRampage/Graphics/GraphicsPipe.h

+++ RoombaRampage/RoombaRampage/Application/ImGui Panels/imgui_handler.cpp
@@ -50,15 +50,15 @@ void ImGuiHandler::Render()
     //create main menu bar
     DrawMainMenuBar();
 
+
+
     Helper::Helpers *help = Helper::Helpers::GetInstance();
-
-
     DrawPerformanceWindow(help->Fps);
     DrawHierachyWindow();
     DrawComponentWindow();
     DrawLogsWindow();
     DrawTestWindow();
-    DrawRenderScreenWindow(Helper::Helpers::GetInstance()->WindowWidth, Helper::Helpers::GetInstance()->WindowHeight);
+    DrawRenderScreenWindow(static_cast<unsigned int>(Helper::Helpers::GetInstance()->WindowWidth), static_cast<unsigned int>(Helper::Helpers::GetInstance()->WindowHeight));
     ImGui::Render();
     ImGuiIO& io = ImGui::GetIO();
     if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
